Refactor error handling
Do not show a top-level error page, as this is particularly distruptive
if we switch sources and network fails: Instead of just keeping the last
image around we now show an error page with no way to get back.

Now we instead just pop up a small toast, with an option to show a more
detailed error dialog.  This is much less disruptive and simplifies
state tracking a lot as we no longer switch stack pages after the
initial image retrieval.

Along the way also remove the separate image download class and instead
inline its properties into the image class which reduces nesting for
a bunch of property bindings.