Download via temporary file
Do not write to the target file directly; instead write to a temporary
file in the target directory, and then rename atomically to the target
file.

This makes sure we never set a partially downloaded file as wallpaper,
because one concurrent download sees the file of another one and assumes
the target already exists.