Add proper navigation bar for image carousel
Move the carousel indicator into an overlay OSD toolbar, and add
navigation buttons.

Add arrow key and page key shortcuts to go to the previous/next image.

Make the navbar appear automatically when these shortcuts are used, or
when the pointer moves in the window. Hide the navbar automatically
after a three seconds, or when the pointer leaves the window.

While adding the navbar, also move outer margins to the main window, for
consistency.

Closes GH-18